Fix all param type

+++ src/resources/B001/retorno/L030/Registro1.php
@@ -178,7 +178,7 @@ class Registro1 extends Generico1
 	* @$data = um array contendo os dados nessesarios para o arquvio
 	*/
 	public function inserirDetalhe($linhaTxt){
-		while($this->data['codigo_lote']==abs(substr(RetornoAbstract::$lines[RetornoAbstract::$linesCounter],3,4)))
+		while($this->data['codigo_lote']==abs((int) substr(RetornoAbstract::$lines[RetornoAbstract::$linesCounter],3,4)))
 		{
 			RetornoAbstract::$linesCounter++;
 			$class = 'CnabPHP\resources\\B'.RetornoAbstract::$banco.'\retorno\\'.RetornoAbstract::$layout.'\Registro3T';
